+++ src/components/common/calendar/DayCalendar.tsx
@@ -5,15 +5,15 @@ import { useRecoilValue, useSetRecoilState } from 'recoil'
 import DateSlideTab from '@/components/main/DateSlideTab'
 import { dateDataType } from '@/components/common/calendar/datePicker/dayDatePIcker'
 import DayDatePicker from '@/components/common/calendar/datePicker/dayDatePIcker'
-import { calendarDateState } from '@/lib/state/calendarDateState'
+import { DayCalendarDateState } from '@/lib/state/calendar/DayCalendarDateState'
 
 import ChevronLeft from '@/icons/icon/chevronLefyt.svg'
 import ChevronRight from '@/icons/icon/chevronRight.svg'
 import CalendarIcon from '@/icons/icon/calendar.svg'
 
 export default function Calendar({ page }: { page: string }) {
-  const calendarDate = useRecoilValue(calendarDateState)
-  const setCalendarDate = useSetRecoilState(calendarDateState)
+  const calendarDate = useRecoilValue(DayCalendarDateState)
+  const setCalendarDate = useSetRecoilState(DayCalendarDateState)
 
   const currentDate = new Date()
   const [dateData, setDateData] = useState<dateDataType>(calendarDate)
