+++ src/lib/state/calendar/DayCalendarDateState.tsx
@@ -1,0 +1,13 @@
+import { atom } from 'recoil'
+import { dateDataType } from '@/components/common/calendar/datePicker/dayDatePIcker'
+
+const date = new Date()
+
+export const DayCalendarDateState = atom<dateDataType>({
+  key: 'dayCalendarDate',
+  default: {
+    year: date.getFullYear(),
+    month: date.getMonth(),
+    date: date.getDate()
+  }
+})
