+++ src/lib/state/calendar/MonthCalendarState.tsx
@@ -1,13 +1,16 @@
 import { atom } from 'recoil'
 import { dateDataType } from '@/components/common/calendar/datePicker/dayDatePIcker'
 
+export interface MonthDateType {
+    year: number;
+    month: number;
+}
 const date = new Date()
 
-export const calendarDateState = atom<dateDataType>({
-  key: 'calendarDate',
+export const MonthCalendarDateState = atom<MonthDateType>({
+  key: 'monthCalendarDate',
   default: {
     year: date.getFullYear(),
     month: date.getMonth(),
-    date: date.getDate()
   }
 })
